requirements.txt
pandas
psycopg2-binary
requests
beautifulsoup4
python-dotenv
ollama
tavily
pypdf
sentence-transformers
numpy
scipy